fix:物件詳細ページを追加
